pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "discord_bot"
dynamic = ["version"]
description = "Discord Bot"
authors = [
    {name = "Tyler D. North", email = "me@tyler-north.com"},
]
requires-python = ">=3.11"
dependencies = [
    "discord.py==2.7.1",
    "pydantic==2.13.1",
    "pydantic-yaml==1.6.0",
    "opentelemetry-sdk==1.41.0",
    # Dont pin these so dependabot updates are easier
    "opentelemetry-instrumentation-requests",
    "opentelemetry-instrumentation-logging",
    "opentelemetry-exporter-otlp",
    "click==8.3.2",
    "pyaml-env==1.2.2",
    "psutil==7.2.2",
    "pytz==2026.1.post1",
    "requests==2.33.1",
]

[project.optional-dependencies]
dispatcher = [
    "redis==7.4.0",
    "async-timeout==5.0.1",
    # Dont pin, so dependabot updates are easier
    "opentelemetry-instrumentation-redis",
]
bot = [
    "async-timeout==5.0.1",
    "aiohttp==3.13.3",
    "dappertable @ git+https://github.com/tnoff/dappertable.git@v1.0.0",
    "opentelemetry-instrumentation-sqlalchemy",
    "alembic==1.18.4",
    "ijson==3.5.0",
    "beautifulsoup4==4.14.3",
    "boto3==1.42.90",
    "google-api-python-client==2.194.0",
    "davey==0.1.5",
    "PyNaCl==1.6.2",
    "SQLAlchemy==2.0.49",
    "spotipy==2.26.0",
    "yt-dlp==2026.3.17",
    # Dont pin, let yt-dlp install
    "yt-dlp-ejs",
    "ytmusicapi==1.11.5",
]
sqlite = [
    "aiosqlite==0.22.1",
]
postgres = [
    "asyncpg==0.31.0",
]
broker = [
    "discord_bot[dispatcher,postgres]",
    "aiohttp==3.13.3",
    "alembic==1.18.4",
    "boto3==1.42.90",
    "SQLAlchemy==2.0.49",
    "opentelemetry-instrumentation-sqlalchemy",
]
all = [
    "discord_bot[dispatcher,bot]",
]
test = [
    "fakeredis==2.35.1",
    "freezegun==1.5.5",
    "pylint-pydantic==0.4.1",
    "pylint==4.0.5",
    "pytest==9.0.3",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.1.0",
    "pytest-freezegun==0.4.2",
    "pytest-mock==3.15.1",
    "requests-mock==1.12.1",
    # Currently required for pytest-freezegun, can remove once they have that fixed
    "setuptools==82.0.1",
    "tox==4.53.0",
]

[project.scripts]
discord-bot = "discord_bot.cli.full:main"
discord-bot-min = "discord_bot.cli.bot:main"
discord-dispatcher = "discord_bot.cli.dispatcher:main"
discord-broker = "discord_bot.cli.broker:main"

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.setuptools.packages.find]
exclude = ["tests*", "docs*"]

[tool.pytest.ini_options]
asyncio_mode = "strict"
log_level = "DEBUG"
asyncio_default_fixture_loop_scope = "function"
python_files = ["tests/*.py"]
filterwarnings = [
    "error",
    "error::requests.exceptions.RequestsDependencyWarning",
    # pytest-freezegun uses the deprecated distutils LooseVersion
    "ignore::DeprecationWarning:pytest_freezegun",
    # discord.py imports audioop (deprecated in 3.11, removed in 3.13); message-based because
    # audioop warns with stacklevel=2 so the warning is attributed to discord/player.py, not audioop
    "ignore:'audioop' is deprecated:DeprecationWarning",
    # discord.py calls asyncio.iscoroutinefunction() (deprecated in 3.14, removed in 3.16);
    # message-based because the warning is attributed to discord/ext/commands/core.py, not asyncio
    "ignore:'asyncio.iscoroutinefunction' is deprecated:DeprecationWarning",
    # music.py uses TemporaryDirectory().name to get a path, immediately discarding the object
    # (intentional pattern, see pylint:disable=consider-using-with comments); stdlib cleans them up
    "ignore:Implicitly cleaning up <TemporaryDirectory:ResourceWarning:tempfile",
    # pytest-asyncio leaks an unclosed event loop between tests in Python 3.11; the two AF_UNIX
    # sockets below are the event loop's internal self-pipe wakeup mechanism and disappear with it;
    # this does not reproduce on Python 3.12+ where event loop creation semantics changed
    "ignore:unclosed event loop:ResourceWarning",
    "ignore:unclosed <socket.socket:ResourceWarning",
]
